Revamp EventFiring capabilities in SeLion.
* Introduced a new Listener named EventListener
which will now provide downstream consumers to tap
into events as and when they happen.
* Introduced a new decorator for CommandExecutor 
implementations that will trigger a before and 
after calls to listeners of type EventListener
* Introduced a new Factory for producing RemoteWebDrivers
or its sub-types such as SelendroidDriver (or)
iOS Driver.
* Introduced sub classes for both SelendroidDriver
and RemoteIOSDriver which automatically inject the 
SeLion decorator for intercepting commands.
* Refactored code across codebase to get away from 
using getWrappedDriver() to access the underlying 
webdriver instances.
* Ripped out all augmenters since they are no longer required.
* Added tests.